Refactor contrast diagnostic
